initial_setup
Reekd the sword and sheath to use as a reference to help with scale.
Basic setup is ready and found some references to help with which
direction I am going to go with this. Seems there are different
variations of the master shield just like the sword. So, this is not
going to be a copy of any one variety however a middle ground that I
will be happy with